rig available for Linux
