+++ src/help-common.in
@@ -6,6 +6,21 @@
 
 use lazy_static::lazy_static;
 
+const HELP_AVAILABLE: &str = r#"
+DESCRIPTION
+    List R versions available to install.
+
+    By default some releases are omitted from the output:
+    - Versions older than R 3.0.0 are omitted. The installation of these
+      might not work at all.
+    - Only the latest release is shown for each minor version. E.g.
+      R 4.2.3 is listed, but other R 4.2.x versions are not.
+    Use `--all` to list all versions.
+
+    Use `--json` to return the output in JSON. JSON output includes the
+    full time stamp (if available) and the download URL as well.
+"#;
+
 const HELP_LIBRARY: &str = r#"
 DESCRIPTION
     Manage package libraries [alias: lib] (experimental)
