+++ src/help-macos.in
@@ -65,21 +65,6 @@ const HELP_RESOLVE_EXAMPLES: &str = r#"EXAMPLES
     rig resolve 4.1
 "#;
 
-const HELP_AVAILABLE: &str = r#"
-DESCRIPTION
-    List R versions available to install.
-
-    By default some releases are omitted from the output:
-    - Versions older than R 3.0.0 are omitted. The installation of these
-      might not work at all.
-    - Only the latest release is shown for each minor version. E.g.
-      R 4.2.3 is listed, but other R 4.2.x versions are not.
-    Use `--all` to list all versions.
-
-    Use `--json` to return the output in JSON. JSON output includes the
-    full time stamp (if available) and the download URL as well.
-"#;
-
 const HELP_ARCH: &str = "Select macOS arch: arm64 or x86_64";
 
 const HELP_DEFAULT: &str = r#"
